node file setup
